requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
build==1.2.1
    # via pip-tools
click==8.1.7
    # via pip-tools
numpy==1.26.4
    # via opencv-python
opencv-python==4.9.0.80
    # via -r requirements.in
packaging==24.0
    # via build
pillow==10.3.0
    # via
    #   -r requirements.in
    #   pillow-heif
pillow-heif==0.16.0
    # via -r requirements.in
pip-tools==7.4.1
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via
    #   build
    #   pip-tools
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
